重构Kindle单词提取工具，新增AI翻译功能，支持增量更新，优化词典条目格式，改进现有单词读取逻辑，增强输出信息的可读性。
